requirements/base.txt
asgiref==3.4.1; python_version >= "3.6"
atomicwrites==1.4.0; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.4.0"
attrs==21.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
black==21.9b0; python_full_version >= "3.6.2"
click==8.0.3; python_version >= "3.6" and python_full_version >= "3.6.2"
colorama==0.4.4; sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.6.2" and platform_system == "Windows" and (python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.5.0")
django-appconf==1.0.5; python_version >= "3.6"
django-environ==0.7.0; python_version >= "3.4" and python_version < "4"
django-imagekit==4.0.2
django-sesame==2.4; python_version >= "3.6" and python_version < "4.0"
django==3.2.8; python_version >= "3.6"
djangorestframework==3.12.4; python_version >= "3.5"
flake8-isort==4.0.0
flake8==3.9.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
iniconfig==1.1.1; python_version >= "3.6"
isort==5.9.3; python_full_version >= "3.6.1" and python_version < "4.0"
mccabe==0.6.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
mypy-extensions==0.4.3; python_full_version >= "3.6.2"
packaging==21.0; python_version >= "3.6"
pathspec==0.9.0; python_full_version >= "3.6.2"
pilkit==2.0
pillow==8.3.2; python_version >= "3.6"
platformdirs==2.4.0; python_version >= "3.6" and python_full_version >= "3.6.2"
pluggy==1.0.0; python_version >= "3.6"
psycopg2-binary==2.9.1; python_version >= "3.6"
py==1.10.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pycodestyle==2.7.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pyflakes==2.3.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
pytest-django==4.4.0; python_version >= "3.5"
pytest==6.2.5; python_version >= "3.6"
pytz==2021.3; python_version >= "3.6"
regex==2021.10.8; python_full_version >= "3.6.2"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
sqlparse==0.4.2; python_version >= "3.6"
testfixtures==6.18.3
toml==0.10.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
tomli==1.2.1; python_version >= "3.6" and python_full_version >= "3.6.2"
typing-extensions==3.10.0.2
